fix: harden JavaScript app execution and resource handling
Interrupt busy scripts on deadlines and reload requests, validate numeric
bindings, and protect widget handles from reuse after deletion. Bound app
timers and replace the growing widget vector and mutex with fixed storage.

Fix Elk string overreads and expired source pointers, remove temporary
allocations from string helpers, preserve SD write bytes, and validate RAM
image sizes and live references before releasing their buffers.

Queue MQTT messages without truncation, expose their topics, restore all
remembered subscriptions, and synchronize runtime command handoffs.
Document limits and recommended JS app patterns.

Validation: ESP32-S3 build with LVGL 8.3.11; native LVGL/Elk regression suite
with address, leak, undefined-behavior, and float-cast-overflow sanitizers.
Physical-device validation remains pending.

+++ webscreen/ws_lvgl_chart_api.h
@@ -149,7 +149,7 @@ static jsval_t js_lv_chart_add_series(struct js *js, jsval_t *args, int nargs) {
   bool available = false;
   for (auto *series : g_chart_series) if (!series) available = true;
   if (!available) return js_mknum(-1);
-  lv_chart_series_t *ser = lv_chart_add_series(obj, lv_color_hex((uint32_t)col), (lv_chart_axis_t)axis);
+  lv_chart_series_t *ser = lv_chart_add_series(obj, lv_color_hex((uint32_t)(int32_t)col), (lv_chart_axis_t)axis);
   if (!ser) return js_mknum(-1);
   // LVGL 8.3 does not initialize the X-buffer ownership flag or its values.
   ser->x_ext_buf_assigned = false;
